mejoras

+++ src/app/borrar-usuario/borrar-usuario.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { UsuarioService } from "../servicios/usuario.service";
 
+import { AuthService } from "../servicios/auth/auth.service";
+
 @Component({
   selector: 'app-borrar-usuario',
   templateUrl: './borrar-usuario.component.html',
@@ -13,8 +15,10 @@ export class BorrarUsuarioComponent implements OnInit {
   @Output()
   stdEvent:EventEmitter<any>;
 
+  token = this.miAuth.getToken();   
 
-  constructor(private perService:UsuarioService)
+
+  constructor(private perService:UsuarioService, private miAuth: AuthService)
   {
     this.stdEvent = new EventEmitter<any>();
   }
@@ -25,14 +29,20 @@ export class BorrarUsuarioComponent implements OnInit {
 
   BorrarUsuario(ID)
   {
-    this.perService.deleteUsuario(ID)
-    .then(data=>{
-
-    console.log(data);
-    this.stdEvent.emit();
-
-    })
-    
+    if (this.token.data.rol == "Admin") {
+      this.perService.deleteUsuario(ID)
+      .then(data=>{
+  
+      console.log(data);
+  
+      this.stdEvent.emit();
+      })
+    }
+    else
+    {
+      console.log("No tiene permitido esa accion");
+      alert("No tiene permitido esa accion");
+    }
   }
